--- README.txt
@@ -1,2 +1,0 @@
-This Inkscape extension is derived from THLaser. It converts paths to gcode for laser cutting. Unlike THLaser, ShinyLaser allows you to set different feedrates for traversal and cut moves. The gcode output is intended to be run on a laser cutter running Smoothieware.
-
