--- shinylaser.inx
@@ -1,58 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<inkscape-extension>
-    <name>ShinyLaser Gcode</name>
-    <id>com.istartup.shinylaser</id>
-
-    <effect>
-        <object-type>path</object-type>
-        <effects-menu>
-            <submenu _name="Export"/>
-        </effects-menu>
-    </effect>
-
-    <dependency type="executable" location="extensions">shinylaser.py</dependency>
-    <dependency type="executable" location="extensions">inkex.py</dependency>
-
-    <script>
-        <command reldir="extensions" interpreter="python">shinylaser.py</command>
-    </script>
-
-    <param name='tab' type="notebook">
-        <page name='tab' _gui-text='Help'>
-            <_param name="fullhelp" type="description">ShinyLaser exports inkscape paths to Gcode compatible with Smoothieware laser module. 
-
-This extenstion is a fork of THLaser v1.30, by Peter Rogers at think|haus, which is itself a fork of Gcodetools v1.2 by Nick Drobchenko.</_param>
-        </page>
-
-        <page name='tab' _gui-text='Advanced'>
-            <param name="biarc-tolerance" type='float' _gui-text='Biarc interpolation tolerance'>1</param>
-            <_param name="help" type="description">Biarc interpolation tolerance is the maximum allowed distance between a path and it's approximation. If this value is exceeded, the path will be split into two segments.</_param>
-            <param name="biarc-max-split-depth" type="int" _gui-text="Maximum splitting depth">4</param>
-            <param name="min-arc-radius" type="float" precision="5" min="-1000" max="1000"  _gui-text="Minimum arc radius">0.05</param>
-            <param name="Xoffset" type="float" precision="4" min="-1000" max="1000" _gui-text="Offset along X axis">0.0</param>
-            <param name="Yoffset" type="float" precision="4" min="-1000" max="1000" _gui-text="Offset along Y axis">0.0</param>
-            <param name="logging" type="boolean" _gui-text="Log debug output from plugin:">true</param>
-        </page>
-        
-        <page name='tab' _gui-text='Preferences'>
-            <param name="filename" type="string" _gui-text="File name: ">output.g</param>
-            <param name="directory" type="string" _gui-text="Directory: "></param>
-            <_param name="help" type="description">(blank is your desktop)</_param>
-            <param name="feed" type="int" min="0" max="1000" _gui-text="Cut Feedrate:">60</param>
-            <param name="Mfeed" type="int" min="0" max="5000" _gui-text="Traversal Feedrate:">300</param>
-            <param name="laser" type="float" precision="2" min="0" max="1" _gui-text="Laser Intensity (0.0-1):">0.1</param>
-            <param name="homebefore" type="boolean" _gui-text="HOMEALL (G28) Before:">true</param>
-            <param name="homeafter" type="boolean" _gui-text="Home X & Y After:">false</param>
-
-            <param name="Xscale" type="float" precision="4" min="-1000" max="1000" _gui-text="Scale along X axis:">1</param>
-            <param name="Yscale" type="float" precision="4" min="-1000" max="1000" _gui-text="Scale along Y axis:">1</param>
-            <param name="double_sided_cutting" type="boolean" _gui-text="Double-sided cutting:">false</param>
-            <param name="draw-curves" type="boolean" _gui-text="Draw curves:">true</param>
-            
-            <param name="unit" type="enum" _gui-text="Units: ">
-                <item value="mm">mm</item>
-                <item value="in">in</item>
-            </param>
-        </page>
-    </param>    
-</inkscape-extension>
