+++ turnkeylaser.inx
@@ -1,0 +1,86 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<inkscape-extension>
+    <name>TurnkeyLaser Exporter</name>
+    <id>com.turnkey.Tyranny</id>
+
+    <effect>
+        <object-type>path</object-type>
+        <effects-menu>
+            <submenu _name="Export"/>
+        </effects-menu>
+    </effect>
+
+    <dependency type="executable" location="extensions">turnkeylaser.py</dependency>
+    <dependency type="executable" location="extensions">inkex.py</dependency>
+
+    <script>
+        <command reldir="extensions" interpreter="python">turnkeylaser.py</command>
+    </script>
+
+    <param name='tab' type="notebook">
+        <page name='tab' _gui-text='Help'>
+            <_param name="fullhelp" type="description">TurnkeyLaser exports Inkscape paths to Gcode compatible with Ramps+Marlin or Smoothieware laser mainboards. 
+            
+You're able to specify in your layer names the laser power output, feedrate and optionally the pulse per millimetre option for pulsed lasing mode as opposed to the default continuous wave operation mode.
+            
+1)Name your layer like the following example : 
+35 [feed=600,ppm=40]
+2)Draw your designs and group them based on lasing options in the layer name.
+3)Select the paths you want to export, then run this script.
+            
+In the example above the layer will be cut at 35% power with a feedrate of 600mm per minute and a pulse rate of 40 pulses per millimetre (at 60ms pulses).
+
+If the ppm option isn't specified in the layer options then output lines will be cut in continuous wave mode at the power and feed specified.
+If you do not specify the laser power or other options in the layer name then they will default to the options in the export dialog box under "Preferences".
+
+For the latest revision please visit https://github.com/TurnkeyTyranny/laser-gcode-exporter-inkscape-plugin
+	    </_param>
+        </page>
+
+        <page name='tab' _gui-text='Advanced'>
+            <param name="mainboard" type="enum" _gui-text="Mainboard: ">
+                <item value="ramps">Ramps 1.4</item>
+                <item value="smoothie">Smoothie</item>
+            </param>
+            <param name="pronterface" type="boolean" _gui-text="Are you using Pronterface?">true</param>
+            <_param name="help" type="description">Will adjust the raster data to be compatible with Pronterface. Very minor loss of laser power on 100% black areas.</_param>
+            
+            <param name="biarc-tolerance" type='float' _gui-text='Biarc interpolation tolerance'>0.5</param>
+            <_param name="help" type="description">Biarc interpolation tolerance is the maximum allowed distance between a path and it's approximation. If this value is exceeded, the path will be split into two segments.</_param>
+            <param name="biarc-max-split-depth" type="int" _gui-text="Maximum splitting depth">4</param>
+            <param name="min-arc-radius" type="float" precision="5" min="-1000" max="5000"  _gui-text="Minimum arc radius">0.00005</param>
+            <param name="Xoffset" type="float" precision="4" min="-1000" max="1000" _gui-text="Offset along X axis">0.0</param>
+            <param name="Yoffset" type="float" precision="4" min="-1000" max="1000" _gui-text="Offset along Y axis">0.0</param>
+            <param name="logging" type="boolean" _gui-text="Log debug output from plugin:">true</param>
+            <param name="optimiseraster" type="boolean" _gui-text="Optimise raster horizontal scanning speed:">true</param>
+            <_param name="help" type="description">Will optimise raster paths, may cause slight overburn at the edges of the raster.</_param>
+            
+        </page>
+        
+        <page name='tab' _gui-text='Preferences'>
+            <param name="origin" type="enum" _gui-text="Y Axis Origin: ">
+                <item value="topleft">Top Left</item>
+                <item value="bottomright">Bottom Left</item>
+            </param>
+            <_param name="help" type="description">Choose the location your laser head parks at home and cuts from</_param>
+            <param name="filename" type="string" _gui-text="File name: ">output.g</param>
+            <param name="directory" type="string" _gui-text="Directory: "></param>
+            <_param name="help" type="description">(blank is your desktop)</_param>
+            <param name="feed" type="int" min="0" max="5000" _gui-text="Default Cut Feedrate:">300</param>
+            <param name="Mfeed" type="int" min="0" max="5000" _gui-text="Default Traversal Feedrate:">2000</param>
+            <param name="laser" type="int" min="0" max="100" _gui-text="Default Laser Intensity (0-100%):">10</param>
+            <param name="homebefore" type="boolean" _gui-text="Home X and Y Before:">true</param>
+            <param name="homeafter" type="boolean" _gui-text="Home X and Y After:">false</param>
+
+            <param name="Xscale" type="float" precision="4" min="-1000" max="1000" _gui-text="Scale along X axis:">1</param>
+            <param name="Yscale" type="float" precision="4" min="-1000" max="1000" _gui-text="Scale along Y axis:">1</param>
+            <param name="double_sided_cutting" type="boolean" _gui-text="Cut both sides of material?">false</param>
+            <param name="draw-curves" type="boolean" _gui-text="Draw curves in inkscape layer">true</param>
+            
+            <param name="unit" type="enum" _gui-text="Units: ">
+                <item value="mm">mm</item>
+                <item value="in">in</item>
+            </param>
+        </page>
+    </param>    
+</inkscape-extension>
